Fer copies de backup

+++ gsx-xarxes/bastio/restaura_inici.sh
@@ -1,20 +1,41 @@
 #!/bin/sh
 
 # lab 1
-cp /etc/network/interfaces.bak /etc/network/interfaces
+if [ -f /etc/network/interfaces.bak ]; then
+	cp /etc/network/interfaces.bak /etc/network/interfaces
+fi
 
 # lab 2
-cp /etc/dhcp/dhcpd.conf.bak /etc/dhcp/dhcpd.conf
-cp /etc/default/isc-dhcp-server.bak /etc/default/isc-dhcp-server
+if [ -f /etc/dhcp/dhcpd.conf.bak ]; then
+	cp /etc/dhcp/dhcpd.conf.bak /etc/dhcp/dhcpd.conf
+fi 
+
+if [ -f /etc/default/isc-dhcp-server.bak ]; then
+	cp /etc/default/isc-dhcp-server.bak /etc/default/isc-dhcp-server
+fi
 
 # lab 3
-cp /etc/bind/named.conf.local.bak /etc/bind/named.conf.local
-cp /etc/bind/named.conf.options.bak /etc/bind/named.conf.options
-cp /etc/resolv.conf.bak /etc/resolv.conf
-cp /etc/hostname.bak /etc/hostname
-cp /etc/hosts.bak /etc/hosts
+if [ -f /etc/bind/named.conf.local.bak ]; then
+	cp /etc/bind/named.conf.local.bak /etc/bind/named.conf.local
+fi
 
-# lab 4
+if [ -f /etc/bind/named.conf.options.bak ]; then
+	cp /etc/bind/named.conf.options.bak /etc/bind/named.conf.options
+fi
+
+if [ -f /etc/resolv.conf.bak ]; then
+	cp /etc/resolv.conf.bak /etc/resolv.conf
+fi
+
+if [ -f /etc/hostname.bak ]; then
+	cp /etc/hostname.bak /etc/hostname
+fi
+
+if [ -f /etc/hosts.bak ]; then
+	cp /etc/hosts.bak /etc/hosts
+fi
 
 # lab 5
-cp /etc/squid/squid.conf.bak /etc/squid/squid.conf
+if [ -f /etc/squid/squid.conf.bak ]; then
+	cp /etc/squid/squid.conf.bak /etc/squid/squid.conf
+fi
